requirements.txt
aiogram==3.13.0
python-dotenv==1.2.1
aiohttp==3.9.5
matplotlib==3.10.8 